Added general configuration page.
A new page has been added to manage general TPS configuration
properties. The properties are read-only by default. In edit
mode the property name will become a link which will show a
dialog to edit the property value.

The config REST service has been updated to use PATCH for
update operation and handle possible null collection of
properties.

Fixed a bug in TableItem.reset() where the code didn't clear
the table cell properly.

Fixed a bug in ConfigDatabase.getProperties() where the code
didn't handle null property key properly.

Ticket #654

+++ base/tps-tomcat/src/org/dogtagpki/server/tps/config/ConfigService.java
@@ -105,15 +105,17 @@ public class ConfigService extends PKIService implements ConfigResource {
             ConfigDatabase configDatabase = new ConfigDatabase();
             ConfigRecord configRecord = configDatabase.getRecord("Generals");
 
-            // validate new properties
             Map<String, String> properties = configData.getProperties();
-            configDatabase.validateProperties(configRecord, null, properties);
+            if (properties != null) {
+                // validate new properties
+                configDatabase.validateProperties(configRecord, null, properties);
 
-            // remove old properties
-            configDatabase.removeProperties(configRecord, null);
+                // remove old properties
+                configDatabase.removeProperties(configRecord, null);
 
-            // add new properties
-            configDatabase.addProperties(configRecord, null, properties);
+                // add new properties
+                configDatabase.addProperties(configRecord, null, properties);
+            }
 
             configDatabase.commit();
